requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.7.0
pillow>=9.5.0
matplotlib>=3.7.0
scikit-image>=0.20.0
tqdm>=4.65.0
scipy>=1.10.0
gradio>=3.35.0
lpips>=0.1.4
